repository/requirements.txt
sqlalchemy>=1.4.41,<2.0
python-dateutil >=2.8.2,<3.0